+++ GameLevel.h
@@ -10,5 +10,5 @@ public:
 	virtual void Load() = 0;
 	virtual void Unload() = 0;
 	virtual void Render() = 0;
-	virtual void Update(long long timeTotal, long long timeDelta) = 0;
+	virtual void Update(double timeTotal, double timeDelta) = 0;
 };
